extract some common prediction logic, make manual predictor use it
