+++ src/config/mod.rs
@@ -1,4 +1,3 @@
-use itertools::Itertools;
 use std::collections::HashMap;
 use std::collections::HashSet;
 use std::error::Error;
@@ -18,7 +17,6 @@ fn match_predictor(predictor: file::Predictor) -> app::Predictor {
             thresholds: thresholds
                 .into_iter()
                 .map(|(k, v)| (k.parse::<u8>().unwrap() as u8, v))
-                .sorted_by_key(|k| k.0)
                 .collect(),
         },
     }
